##11 typos and escaping some 's
In a couple of places it said `egrep`s followed by a second code like `STDIN` which messed things up. I changed it to `egrep\'s` and now it looks much better. If the 's is successfully outside the code, it's HUGE and looks real dumb next to the code.